==> notebooks/02_eda_microbus.ipynb <==
# EDA for theft inside **microbus** units in CDMX
This notebook explores how many rows of data can be extracted for a data subset that focuses on the issue of thefts in public transportation units called **microbuses** or **peseros** reported for CDMX. This is a insecurity topic that the city has not been able to control yet

import pandas as pd
import zipfile
import numpy as np
import matplotlib.pyplot as plt

#Path
#Insert your local repo path to file 
repo_path = "INSERT YOUR LOCAL SDC-SECURITY REPO PATH HERE"
repo_path = "D:\Archivos\Social Data Challenge\sdc-security"

crimes= pd.read_csv(repo_path+'\datasets\crimes.csv')

crimes.head()

categorias = crimes.categoria.unique()
print(categorias)

microbus=crimes[crimes["categoria"] == "ROBO A PASAJERO A BORDO DE MICROBUS CON Y SIN VIOLENCIA"]

print(microbus.shape)
microbus.head()

crimes.categoria.unique()
