==> notebooks/03_colonia_homologation.ipynb <==
# Homologation of **colonias**

The purpose of this notebook is to homologate the *colonia* values in the *crimes_clean* dataset using a JSON file called *catalogo-de-colonias.json*, which includes names and keys of the location in detail.

import pandas as pd
import zipfile
import numpy as np
import matplotlib.pyplot as plt
import json

#Path
#Insert your local repo path to file 
repo_path = "INSERT YOUR LOCAL SDC-SECURITY REPO PATH HERE"
repo_path ='D:\Archivos\Social Data Challenge\sdc-security'

crime_data_path = "\datasets\crimes_clean.csv.zip"

zf = zipfile.ZipFile(repo_path+crime_data_path) 
crimes = pd.read_csv(zf.open('crimes_clean.csv'))
print(crimes.shape)


#Changing categorical variable 'sexo' to numerical

#Non optimal way of doing this is using lambda functions
#crimes["sexo_fem"] = crimes['sexo'].apply(lambda x: 1 if x.upper() == 'FEMENINO' else 0)

#np.where uses vectorized operations which are highly optimized
if 'sexo' in crimes.columns:
    crimes['sexo_fem'] = np.where(crimes['sexo'].str.upper() == 'FEMENINO', 1, 0)
    del crimes["sexo"]

crimes.head()


### Extracting info from JSON File 'catalogo-de-colonias.json' 

catalog_path = "\datasets\catalogo-de-colonias.json"

# Load the JSON file
with open(repo_path + catalog_path, 'r', encoding='utf-8') as json_file:
    data = json.load(json_file)

# Extracting features from the JSON data
features = data['features']

# Initialize lists to store column values
polygon_list = []
properties_list = []

# Extracting values from features
for feature in features:
    polygon_list.append(feature['geometry']['coordinates'][0])  # Extracting the first set of coordinates
    properties_list.append(feature['properties'])

# Creating the DataFrame
catalog = pd.DataFrame(properties_list)
catalog['polygon'] = polygon_list

# Reordering columns
column_order = ['cve_col', 'clasif', 'colonia', 'alc', 'cve_alc', 'cve_ent', 'entidad', 'polygon']
catalog= catalog[column_order]


# Display the DataFrame
catalog.head()


Reviewing how does a value inside 'polygon' looks like


catalog["polygon"].loc[0]

print(catalog.clasif.unique())
print(catalog.entidad.unique())
print(catalog.cve_ent.unique())

Erasing cve_ent and entidad, since it is the same on the whole catalog


del catalog["cve_ent"]
del catalog["entidad"]

catalog.head()

After erasing some columns that will not be used, spanish accented character should be replaced by their non-accented counterpart

pip install unidecode


from unidecode import unidecode


catalog_cat_cols = ['clasif', 'colonia', 'alc']

crimes_cat_cols  = ['delito', 'categoria', 'alcaldia', 'colonia',
                    'colonia_alt', 'tipopersona', 'calidadjuridica',
                    'competencia']

#Dataframes and columns to be processed in dictionary
categorical_dict = {'catalog': catalog_cat_cols, 'crimes':  crimes_cat_cols}

for df_name, columns in categorical_dict.items():
    df = globals()[df_name]
    for col in columns:
        df[col] = df[col].apply(unidecode)
        df[col] = df[col].str.upper()

catalog.head()

Focusing on the CUAUHTEMOC district, a slice of this catalog containg only information related to that district will be created


catalog_cuau = catalog.loc[ catalog["alc"] == 'CUAUHTEMOC'].copy()
catalog_cuau.head()

# How many colonias are related to Cuauhtemoc district on this catalog?
catalog_cuau.colonia.unique().shape[0]

The next step is to homologate the *colonia* names in the *crimes* dataset and fix the names when needed so that, in future merges with other datasets, the catalog is used as the naming convention.

# Reordering columns
column_order = ['idcarpeta', 'delito', 'categoria', 'alcaldia', 'colonia', 'sexo_fem',
                'edad', 'tipopersona', 'calidadjuridica', 'anio_denuncia', 
                'mes_denuncia', 'fecha_denuncia', 'hora_denuncia', 'competencia', 
                'anio_hecho','mes_hecho', 'fecha_hecho', 'hora_hecho', 'colonia_alt', 
                'crimen_lat', 'crimen_lon', 'hecho_time', 'denuncia_time']
crimes= crimes[column_order]
crimes.head()


crimes_cuau = crimes[crimes.alcaldia == 'CUAUHTEMOC'].copy()

Checking how many rows are on each dataframe and how many are left when a merge is done

print('Crimes dataframe before merge contains {} rows'.format(crimes_cuau.shape[0]))
print(crimes_cuau.isnull().sum())

print('Catalog dataframe before merge contains {} rows'.format(catalog_cuau.shape[0]))
print(catalog_cuau.isnull().sum())

merge_colonias_cuau = crimes_cuau.merge(catalog_cuau, left_on = ['alcaldia', 'colonia'],
                                 right_on = ['alc','colonia'], how = 'left')

print('Merge dataframe contains {} rows'.format(merge_colonias_cuau.shape[0]))
print(merge_colonias_cuau.isnull().sum())
merge_colonias_cuau.head()

The following code block uses the rows with null values to identify which *colonias* values need to be homologated with the *catalogo-de-colonias.json* so that a successful merge is done after that.

By checking values inside the *catalogo-de-colonias.json* and validating with the *crimes.colonia_alt* column, a dictionary with substitute *colonia* values is presented. 


condition = merge_colonias_cuau.cve_col.isnull()
missing_colonias = list(merge_colonias_cuau.loc[condition,'colonia'].unique())
missing_colonias
#merge_colonias_cuau[merge_colonias_cuau.cve_col.isnull()]

The substitutions will be made from the *crimes_clean* dataframe in case a future analysis with other alcaldias is needed. Because of this, the merge with the *catalogo* dataframe will be done again.

A lot of inconsistencies happen because of the roman numerals, so these will be removed.

Checking the number of ocurrences of each missing colonia value. If the number is low and it does not contain roman numerals, it is possible that the *colonia* actually belongs to other *alcaldia*

for colonia in missing_colonias:
    print(colonia + ': '+ str(crimes_cuau[crimes_cuau.colonia == colonia].shape))


While checking the coordinates in dataset, it was noticed that there are rows which contain the same values of crimen_lat and crimen_lon. This rows will be deleted for now, since they wont be useful for geographical location

coordinates_condition = crimes.crimen_lat == crimes.crimen_lon 
print(crimes.loc[coordinates_condition].shape)

print('With those rows, shape is ',crimes.shape)
crimes=crimes[~coordinates_condition]
print('Without those rows, shape is ',crimes.shape)

A validation was made using the crime_lat and crime_long columns and Google Maps so that, for each case, it is known if the *colonia* is the one to be fixed or the *alcaldia* column

homologation_dict = {
    'AMPLIACION ASTURIAS': 'AMPL. ASTURIAS',
    'CENTRO URBANO BENITO JUAREZ': 'ROMA SUR',
    'EX-HIPODROMO DE PERALVILLO': 'EX HIPODROMO DE PERALVILLO',
    'LORENZO BOTURINI':'TRANSITO',
    'MERCED BALBUENA': 'ESPERANZA',
    'SAN MIGUEL CHAPULTEPEC':'CONDESA',
    'ZONA CENTRO': 'CENTRO'
}

alcaldia_homologation = {
    'AGRICULTURA': 'MIGUEL HIDALGO',
    'AMPLIACION DEL GAS':'AZCAPOTZALCO',
    'NICOLAS BRAVO': 'VENUSTIANO CARRANZA',
    'TLATILCO': 'AZCAPOTZALCO',
    'VALLEJO': 'GUSTAVO A. MADERO',
}



import re
# Function to remove Roman numerals
def remove_roman_numerals(text):
    return re.sub(r'\s*\b[IVXLCDM]+\b\s*', '', text)

condition =  crimes['alcaldia'] == 'CUAUHTEMOC'
condition2 = crimes['colonia'].isin(missing_colonias) 

crimes.loc[condition & condition2, 'colonia'] = crimes['colonia'].apply(remove_roman_numerals)
crimes.loc[condition, 'colonia'] = crimes.loc[condition, 'colonia'].replace(homologation_dict)
crimes['alcaldia'] = crimes['colonia'].map(alcaldia_homologation).fillna(crimes['alcaldia'])


crimes_cuau = crimes[crimes.alcaldia == 'CUAUHTEMOC'].copy()

Validating that all the rows were merged succesfully without null values

print('Crimes dataframe before merge contains {} rows'.format(crimes_cuau.shape[0]))
print(crimes_cuau.isnull().sum())

print('Catalog dataframe before merge contains {} rows'.format(catalog_cuau.shape[0]))
print(catalog_cuau.isnull().sum())

merge_colonias_cuau = crimes_cuau.merge(catalog_cuau, left_on = ['alcaldia', 'colonia'],
                                 right_on = ['alc','colonia'], how = 'left')

print('Merge dataframe contains {} rows'.format(merge_colonias_cuau.shape[0]))
print(merge_colonias_cuau.isnull().sum())


condition = merge_colonias_cuau.cve_col.isnull()
missing_colonias = list(merge_colonias_cuau.loc[condition,'colonia'].unique())
missing_colonias

crimes.to_csv('../datasets/crimes_clean2.csv.zip', compression = 'zip', index=False)


merge_colonias_cuau.to_csv('../datasets/cuau_homologated.csv.zip', compression = 'zip', index=False)

catalog.to_csv('../datasets/catalogo_colonias.csv.zip', compression = 'zip', index=False)
