==> notebooks/06_from_s3.ipynb <==
!pip install pandas boto3
import pandas as pd
import boto3
import botocore
from io import StringIO

import zipfile




# Create an S3 client

bucket_name = 'sdc-security'
crimes_path = 'datasets/crimes-cams.csv'
density_path = 'datasets/densidad_poblacion_2020.csv'

local_crimes_path = "D:\Archivos\Social Data Challenge\QGIS\crimes-cams.csv"
local_density_path = "D:\Archivos\Social Data Challenge\data\densidad_poblacion_2020.csv"


def s3_bucket_csv_to_pd(bucket_name, file_path):
    
    try:
        # Create an S3 client
        s3 = boto3.client('s3')

        # Download the CSV file from S3
        s3_object = s3.get_object(Bucket=bucket_name, Key=file_path)

        # Read the CSV content
        csv_content = s3_object['Body'].read().decode('utf-8')

        # Create a DataFrame from the CSV content
        df = pd.read_csv(StringIO(csv_content))

        return df
    except botocore.exceptions.NoCredentialsError:
        # Handle the case where AWS credentials are missing or invalid
        print("AWS credentials are missing or invalid.")
        return None
    except botocore.exceptions.ClientError as e:
        # Handle specific S3 client errors
        if e.response['Error']['Code'] == 'NoSuchKey':
            # Handle the case where the S3 file does not exist
            print("S3 file does not exist.")
            return None
        else:
            # Handle other S3 client errors
            print("Error reading from S3:", e)
            return None
    except Exception as e:
        # Handle other exceptions
        print("Error:", e)
        return None


def load_dataframe(bucket, s3_file_path, local_file_path):
    df = s3_bucket_csv_to_pd(bucket, s3_file_path)
    
    if df is None:
        print(f"Reading DataFrame from local file: {local_file_path}")
        df = pd.read_csv(local_file_path)
    return df

gsi_crimes = load_dataframe(bucket_name, crimes_path, local_crimes_path)
density    = load_dataframe(bucket_name, density_path, local_density_path)
 

density.rename(columns = lambda x : x.lower() , inplace = True)
density = density[density.alcaldia == 'CUAUHTEMOC'] 

catalogo_path = "\datasets\catalogo_colonias.csv.zip"
repo_path ='D:\Archivos\Social Data Challenge\sdc-security'
zf = zipfile.ZipFile(repo_path+catalogo_path) 
catalogo= pd.read_csv(zf.open('catalogo_colonias.csv'))
print(catalogo.shape)

catalogo_cuau = catalogo[catalogo.alc == 'CUAUHTEMOC']

catalogo_cuau.colonia.unique()

homologation_dict = {
    'AMPL. ASTURIAS':['ASTURIAS (AMPL)'], 
    'BUENAVISTA':['BUENAVISTA I', 'BUENAVISTA II'], 
    'CENTRO':['CENTRO I','CENTRO II', 
              'CENTRO III', 'CENTRO IV', 
              'CENTRO V', 'CENTRO VI',
              'CENTRO VII', 'CENTRO VIII'],
    'DOCTORES':['DOCTORES I', 'DOCTORES II', 
                'DOCTORES III', 'DOCTORES IV', 'DOCTORES V'],
    'GUERRERO':['GUERRERO I', 'GUERRERO II', 
                'GUERRERO III', 'GUERRERO IV'],
    'HIPODROMO':['HIPODROMO I', 'HIPODROMO II'], 
    'MORELOS':['MORELOS I', 'MORELOS II', 'MORELOS III'],
    'NONOALCO TLATELOLCO':['NONOALCO-TLATELOLCO (U HAB) I', 
                           'NONOALCO-TLATELOLCO (U HAB) II',
                           'NONOALCO-TLATELOLCO (U HAB) III',
                           'NONOALCO-TLATELOLCO'], 
    'OBRERA': ['OBRERA I', 'OBRERA II','OBRERA III', 'OBRERA IV'], 
    'PERALVILLO': ['PERALVILLO I', 'PERALVILLO II'], 
    'ROMA NORTE': ['ROMA NORTE I', 'ROMA NORTE II', 'ROMA NORTE III'],
    'ROMA SUR': ['ROMA SUR I', 'ROMA SUR II'], 
    'SAN RAFAEL':['SAN RAFAEL I', 'SAN RAFAEL II'],
    'SANTA MARIA LA RIBERA':['SANTA MARIA LA RIBERA I', 
                             'SANTA MARIA LA RIBERA II',
                             'SANTA MARIA LA RIBERA IV'], 
    'SANTA MARIA INSURGENTES':['SANTA MARIA (U HAB)'],
    'ATLAMPA':[],
    'ALGARIN' :[], 
    'ASTURIAS':[], 
    'BUENOS AIRES':[], 
    'CONDESA':[], 
    'CUAUHTEMOC':[], 
    'ESPERANZA':[], 
    'EX HIPODROMO DE PERALVILLO':[], 
    'FELIPE PESCADOR':[],
    'HIPODROMO CONDESA':[], 
    'JUAREZ':[],
    'MAZA':[], 
    'PAULINO NAVARRO':[], 
    'SAN SIMON TOLNAHUAC':[], 
    'TABACALERA':[], 
    'TRANSITO':[],
    'VALLE GOMEZ':[], 
    'VISTA ALEGRE':[], 
}



# Create 'colonia_new' column by mapping values from the dictionary
density['colonia_new'] = density['colonia'].apply(
    lambda x: next((k for k, v in homologation_dict.items() if x in v), x))

sum_aggregation_columns = ['area_ha','pobtot2010', 'pobtot2020', 
                           'vivhab2010', 'vivhab2020']

for col in sum_aggregation_columns:
    # Create aggregated columns by sum
    density[f'{col}_new'] = density.groupby('colonia_new')[col].transform('sum')
    
    # Create columns related to density
    if col not in ['area_ha']:
        #{col[0:3]} gets 'pob' or 'viv'
        #{col[-4:]} gets '2010' or '2020'
        density[f'den{col[0:3]}{col[-4:]}'] = density[f'{col}_new'] / density['area_ha_new']




other_cols_to_drop = ['fid_1','cve_ent','alcaldia','cve_col','denpob10','denpob20','denviv10','denviv20','difdenpob','difdenviv','tcma10_20']

density.drop(columns=['colonia'] + sum_aggregation_columns + other_cols_to_drop, 
             inplace=True)


density.drop_duplicates(subset=['colonia_new'], inplace=True)
density.columns = density.columns.to_series().apply(lambda x: x.replace('_new', ''))
density.columns = density.columns.to_series().apply(lambda x: x.replace('tot', ''))
density[density.colonia =='DOCTORES']

density['dif_denpob'] = density['denpob2020'] - density['denpob2010'] 

density = density[['colonia','pob2010', 'pob2020', 'denpob2010','denpob2020', 'dif_denpob',
                  'area_ha','vivhab2010', 'vivhab2020','denviv2010','denviv2020']]


density.shape

print(gsi_crimes.isnull().sum())
crimes_cams_density = pd.merge(gsi_crimes,density, on = ['colonia'], how = 'left')
print(crimes_cams_density.isnull().sum())
crimes_cams_density.head()

crimes_cams_density['crimes_in_col'] = crimes_cams_density.groupby('colonia')['colonia'].transform('count')
crimes_cams_density['den_crime'] = crimes_cams_density['crimes_in_col'] / crimes_cams_density['area_ha']



crimes_cams_density.head()

### Merging with total number of C5 cams on each colonia 

c5_cams_cuau = pd.read_csv('D:\Archivos\Social Data Challenge\QGIS\c5_cuau.csv')
c5_cams_cuau.colonia.unique()

catalogo_cuau.colonia.unique()

c5_agg = c5_cams_cuau.groupby('colonia').agg(count=('colonia','count'))
#c5_agg = consolidated.groupby('colonia').agg(count=('colonia','count'), alcaldia=('alcaldia', 'first'))
c5_agg.reset_index(inplace=True)
c5_agg.rename(columns= {'count':'cams_col_count'}, inplace = True)

crimes_cams_density=crimes_cams_density.merge(c5_agg, on =['colonia'], how = 'left')

crimes_cams_density['cams_density'] = crimes_cams_density['cams_col_count']/ crimes_cams_density['area_ha']
crimes_cams_density['crimes_cams_den'] = crimes_cams_density['crimes_in_col'] / crimes_cams_density['cams_col_count']
crimes_cams_density.head()


crimes_cams_density.to_csv('../datasets/crimes_cams_density.csv', index=False)
density.to_csv('../datasets/density_cuau.csv', index=False)
