==> notebooks/07_merging_qgis_data.ipynb <==
import pandas as pd
import zipfile
import numpy as np
import matplotlib.pyplot as plt
import json
import os


#Path
#Insert your local repo path to file
path_home = "D:\Archivos\Social Data Challenge\data\grids"
path_500 = path_home + "\g500"
path_100 = path_home + "\g100"


csv_100 = [file for file in os.listdir(path_100) if file.endswith('.csv')]
csv_500 = [file for file in os.listdir(path_500) if file.endswith('.csv')]


#TODO: refactor to avcoid duplicating this code for each grid
dataframes = []


for file in csv_500:
    df = pd.read_csv(os.path.join(path_500, file))
    last_column_name = df.columns[-1]
    df = df[['id', 'lat', 'lon', last_column_name]]
    dataframes.append(df)

#merged_df = dataframes[0]  # Start with the first DataFrame
merged_df = pd.read_csv(path_home +'\g500_2019_quarters_coordinates.csv')
merged_df=merged_df.drop(columns = ['left','top','right','bottom'])
for df in dataframes:
    merged_df = pd.merge(merged_df, df, on=['id', 'lat', 'lon'], how='inner')
    

grid_500 = merged_df.copy()

grid_500.shape


dataframes = []

for file in csv_100:
    df = pd.read_csv(os.path.join(path_100, file))
    last_column_name = df.columns[-1]
    df = df[['id', 'lat', 'lon', last_column_name]]
    dataframes.append(df)

#merged_df = dataframes[0]  # Start with the first DataFrame
merged_df = pd.read_csv(path_home +'\g100_2019_quarters_coordinates.csv')
merged_df=merged_df.drop(columns = ['left','top','right','bottom'])
for df in dataframes:
    merged_df = pd.merge(merged_df, df, on=['id', 'lat', 'lon'], how='inner')
    
    

grid_100 = merged_df.copy()

grid_100.shape

### Grouping 'zonas de patrullaje' grid data 

patrullaje_500_path = 'D:\Archivos\Social Data Challenge\data\grids\zonas_patrullaje_500.csv'
patrullaje_100_path = 'D:\Archivos\Social Data Challenge\data\grids\zonas_patrullaje_100.csv'

patrullaje_500 = pd.read_csv(patrullaje_500_path)
patrullaje_100 = pd.read_csv(patrullaje_100_path)

pat_500 = patrullaje_500.groupby(by=['id','lat','lon']).size().reset_index(name='zonas_pat')
pat_100 = patrullaje_100.groupby(by=['id','lat','lon']).size().reset_index(name='zonas_pat')


grid_100 = pd.merge(grid_100, pat_100, on=['id', 'lat', 'lon'], how='inner')
grid_500 = pd.merge(grid_500, pat_500, on=['id', 'lat', 'lon'], how='inner')

grid_500.columns


grid_100.columns


grid_500 = grid_500[['id', 'lat', 'lon', 'metro_500', 'cams_500', 'zonas_pat',
                     'antro_500', 'salud_500', 'museo_500', 'escuela_50', 'deporte_50',
                     'g500_19q1', 'g500_19q2', 'g500_19q3', 'g500_19q4',      
                     'g500_20q1', 'g500_20q2', 'g500_20q3', 'g500_20q4', 
                     'g500_21q1', 'g500_21q2', 'g500_21q3', 'g500_21q4',
                     'g500_22q1', 'g500_22q2', 'g500_22q3', 'g500_22q4', 
                     'g500_23q1'
                    ]]

grid_100 = grid_100[[ 'id', 'lat', 'lon', 'metro_100', 'cams_100', 'zonas_pat',
                      'antro_100', 'salud_100', 'museo_100', 'escuela_10', 'deporte_10',
                      'g100_19q1', 'g100_19q2', 'g100_19q3', 'g100_19q4',        
                      'g100_20q1', 'g100_20q2', 'g100_20q3', 'g100_20q4', 
                      'g100_21q1', 'g100_21q2', 'g100_21q3', 'g100_21q4',
                      'g100_22q1', 'g100_22q2', 'g100_22q3', 'g100_22q4', 
                      'g100_23q1',
                      ]]

grid_500.rename(columns ={'escuela_50':'esc_500','deporte_50':'depo_500'}, inplace =True)
grid_100.rename(columns ={'escuela_10':'esc_100','deporte_10':'depo_100'}, inplace =True)

grid_100.isnull().sum()

#Validation lines

print(grid_100.shape)
print(grid_500.shape)

print(grid_100.isnull().sum())
print(grid_500.isnull().sum())


grid_100.columns

grid_500.to_csv('../datasets/grid_500.csv', index = False)
grid_100.to_csv('../datasets/grid_100.csv', index = False)

# Joining Heatmap samples info


path_home = "D:\Archivos\Social Data Challenge\data\grids"

#csv_100 = pd.read_csv(path_home + '\grid_100.csv')

path_100 = path_home + "\quarter_samples\grid100"
path_500 = path_home + "\quarter_samples\grid500"

samples_100 = [file for file in os.listdir(path_100) if file.endswith('.csv')]
samples_500 = [file for file in os.listdir(path_500) if file.endswith('.csv')]

dataframes = []

for file in samples_500:
    df = pd.read_csv(os.path.join(path_500, file))
    last_column_name = df.columns[-1]
    df = df[['id', 'lat', 'lon', last_column_name]]
    dataframes.append(df)

#merged_df = pd.read_csv(path_home + '\grid_500.csv')
merged_df = grid_500.copy()

for df in dataframes:
    merged_df = pd.merge(merged_df, df, on=['id', 'lat', 'lon'], how='left')
    
grid_500_htmp = merged_df.copy()
print(grid_500_htmp.shape)
print(grid_500_htmp.isnull().sum())
grid_500_htmp.head()


dataframes = []

for file in samples_100:
    df = pd.read_csv(os.path.join(path_100, file))
    last_column_name = df.columns[-1]
    df = df[['id', 'lat', 'lon', last_column_name]]
    dataframes.append(df)

#merged_df = pd.read_csv(path_home + '\grid_100.csv')
merged_df = grid_100.copy()

for df in dataframes:
    merged_df = pd.merge(merged_df, df, on=['id', 'lat', 'lon'], how='left')
    
grid_100_htmp = merged_df.copy()
print(grid_100_htmp.shape)
print(grid_100_htmp.isnull().sum())
grid_100_htmp.head()


for df in [grid_100_htmp, grid_500_htmp]:
    df.fillna(0,inplace=True)
    df.columns = df.columns.str.replace('g100_', 'crime')
    df.columns = df.columns.str.replace('g500_', 'crime')


grid_100.columns

grid_100_htmp= grid_100_htmp[[  'id', 'lat', 'lon', 
                                'metro_100', 'cams_100', 'zonas_pat', 
                                'antro_100', 'salud_100', 'museo_100', 'esc_100', 'depo_100', 
                                'crime19q1', 'sample19q1', 'crime19q2', 'sample19q2', 
                                'crime19q3', 'sample19q3', 'crime19q4', 'sample19q4',
                                'crime20q1', 'sample20q1', 'crime20q2', 'sample20q2', 
                                'crime20q3', 'sample20q3', 'crime20q4', 'sample20q4',
                                'crime21q1', 'sample21q1', 'crime21q2', 'sample21q2',
                                'crime21q3', 'sample21q3', 'crime21q4', 'sample21q4',
                                'crime22q1', 'sample22q1', 'crime22q2', 'sample22q2', 
                                'crime22q3', 'sample22q3', 'crime22q4', 'sample22q4',
                                'crime23q1', 'sample23q1']]

grid_500_htmp= grid_500_htmp[[  'id', 'lat', 'lon', 
                                'metro_500', 'cams_500', 'zonas_pat', 
                                'antro_500', 'salud_500',  'museo_500', 'esc_500', 'depo_500', 
                                'crime19q1', 'sample19q1', 'crime19q2', 'sample19q2', 
                                'crime19q3', 'sample19q3', 'crime19q4', 'sample19q4',
                                'crime20q1', 'sample20q1', 'crime20q2', 'sample20q2', 
                                'crime20q3', 'sample20q3', 'crime20q4', 'sample20q4',
                                'crime21q1', 'sample21q1', 'crime21q2', 'sample21q2',
                                'crime21q3', 'sample21q3', 'crime21q4', 'sample21q4',
                                'crime22q1', 'sample22q1', 'crime22q2', 'sample22q2', 
                                'crime22q3', 'sample22q3', 'crime22q4', 'sample22q4',
                                'crime23q1', 'sample23q1']]

grid_100_htmp

grid_500.to_csv('../datasets/grid_500_htmp.csv', index = False)
grid_100.to_csv('../datasets/grid_100_htmp.csv', index = False)
