==> notebooks/13_mergeFolder_GridSearch.ipynb <==
import pandas as pd
import zipfile
import numpy as np
import matplotlib.pyplot as plt
import json
import os


from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import StandardScaler
from sklearn.feature_selection import mutual_info_classif
from sklearn import tree
from sklearn.model_selection import train_test_split
from sklearn.metrics import (
		accuracy_score as acc_score,
		recall_score   as recall,
		confusion_matrix,
        mean_squared_error as mse, 
        r2_score
		)



#Path
#Insert your local repo path to file
path_home = "D:\Archivos\Social Data Challenge\data\grids"

csv_100 = pd.read_csv(path_home + '\grid_100.csv')

path_100 = path_home + "\quarter_samples\grid100"
path_500 = path_home + "\quarter_samples\grid500"

samples_100 = [file for file in os.listdir(path_100) if file.endswith('.csv')]
samples_500 = [file for file in os.listdir(path_500) if file.endswith('.csv')]




test100= pd.read_csv(path_home + '\grid_100.csv')
print(test100.shape)
print(test100.isnull().sum())

test500= pd.read_csv(path_home + '\grid_500.csv')
print(test500.shape)
print(test500.isnull().sum())

This will read each CSV file from a specific folder, retrieve the last column from each file and merge it into the main dataframe

dataframes = []

for file in samples_500:
    df = pd.read_csv(os.path.join(path_500, file))
    last_column_name = df.columns[-1]
    df = df[['id', 'lat', 'lon', last_column_name]]
    dataframes.append(df)

merged_df = pd.read_csv(path_home + '\grid_500.csv')
for df in dataframes:
    merged_df = pd.merge(merged_df, df, on=['id', 'lat', 'lon'], how='left')
    
grid_500_htmp = merged_df.copy()
print(grid_500_htmp.shape)
print(grid_500_htmp.isnull().sum())
grid_500_htmp.head()


dataframes = []

for file in samples_100:
    df = pd.read_csv(os.path.join(path_100, file))
    last_column_name = df.columns[-1]
    df = df[['id', 'lat', 'lon', last_column_name]]
    dataframes.append(df)

merged_df = pd.read_csv(path_home + '\grid_100.csv')
for df in dataframes:
    merged_df = pd.merge(merged_df, df, on=['id', 'lat', 'lon'], how='left')
    
grid_100_htmp = merged_df.copy()
print(grid_100_htmp.shape)
print(grid_100_htmp.isnull().sum())
grid_100_htmp.head()


for df in [grid_100_htmp, grid_500_htmp]:
    df.fillna(0,inplace=True)
    df.columns = df.columns.str.replace('g100_', 'crime')
    df.columns = df.columns.str.replace('g500_', 'crime')



grid_100_htmp.columns


grid_100_htmp= grid_100_htmp[[  'id', 'lat', 'lon', 
                                'metro_100', 'cams_100', 'zonas_pat', 
                                'antro_100', 'salud_100', 'museo_100', 'esc_100', 'depo_100', 
                                'crime19q1', 'sample19q1', 'crime19q2', 'sample19q2', 
                                'crime19q3', 'sample19q3', 'crime19q4', 'sample19q4',
                                'crime20q1', 'sample20q1', 'crime20q2', 'sample20q2', 
                                'crime20q3', 'sample20q3', 'crime20q4', 'sample20q4',
                                'crime21q1', 'sample21q1', 'crime21q2', 'sample21q2',
                                'crime21q3', 'sample21q3', 'crime21q4', 'sample21q4',
                                'crime22q1', 'sample22q1', 'crime22q2', 'sample22q2', 
                                'crime22q3', 'sample22q3', 'crime22q4', 'sample22q4',
                                'crime23q1', 'sample23q1']]

grid_500_htmp= grid_500_htmp[[  'id', 'lat', 'lon', 
                                'metro_500', 'cams_500', 'zonas_pat', 
                                'antro_500', 'salud_500',  'museo_500', 'esc_500', 'depo_500', 
                                'crime19q1', 'sample19q1', 'crime19q2', 'sample19q2', 
                                'crime19q3', 'sample19q3', 'crime19q4', 'sample19q4',
                                'crime20q1', 'sample20q1', 'crime20q2', 'sample20q2', 
                                'crime20q3', 'sample20q3', 'crime20q4', 'sample20q4',
                                'crime21q1', 'sample21q1', 'crime21q2', 'sample21q2',
                                'crime21q3', 'sample21q3', 'crime21q4', 'sample21q4',
                                'crime22q1', 'sample22q1', 'crime22q2', 'sample22q2', 
                                'crime22q3', 'sample22q3', 'crime22q4', 'sample22q4',
                                'crime23q1', 'sample23q1']]

grid_100_htmp

grid_100_htmp.to_csv('../datasets/grid_100_htmp.csv', index = False)
grid_500_htmp.to_csv('../datasets/grid_500_htmp.csv', index = False)


grid_100_htmp.head()

# Machine Learning Benchmarking

#MODELS
from sklearn.linear_model import LinearRegression
from sklearn import tree
from sklearn.linear_model import Perceptron
from sklearn.tree import DecisionTreeRegressor
from sklearn.linear_model import LinearRegression
from sklearn.neighbors import KNeighborsRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.svm import SVR
#import xgboost as xgb
#xgb_regressor = xgb.XGBRegressor()
from sklearn.neural_network import MLPRegressor  # Use MLPRegressor for regression
#from minepy import MINE

from sklearn.decomposition import PCA


X =  grid_100_htmp.iloc[:, 1:-2]
y_crime23q1 = grid_100_htmp['crime23q1']
y_sample23q1 = grid_100_htmp['sample23q1']

scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)
x_1=pd.DataFrame(X_scaled, columns = X.columns)
#X_scaled[X_scaled.sample22q2 != 0]
#x[x.crime22q4 >0]

pca = PCA(n_components=2)
pca.fit(x_1)
x_pca=pca.transform(x_1)

#crime23q1_predictions

# Grid Search

from sklearn.model_selection import GridSearchCV


X =  grid_100_htmp.iloc[:, 1:-2]
y_crime23q1 = grid_100_htmp['crime23q1']
y_sample23q1 = grid_100_htmp['sample23q1']

#scaler = StandardScaler()
#X_scaled = scaler.fit_transform(X)
#x_1=pd.DataFrame(X_scaled, columns = X.columns)

#X_scaled[X_scaled.sample22q2 != 0]
#x[x.crime22q4 >0]

# Split the dataset into training and testing sets

y = y_sample23q1
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)


models = {
    'Linear Regression': LinearRegression(),
    'Random Forest Regressor': RandomForestRegressor(),
    'Support Vector Regressor': SVR(),
    'Perceptron': Perceptron(),
    'Neural Network': MLPRegressor()
    # Add more regressor models and hyperparameters here
}

# Define a parameter grid for each model
param_grids = {
    'Linear Regression': {},
    'Random Forest Regressor': {
        'n_estimators': [100, 200],
        'max_depth': [None, 10, 20],
        # Add more hyperparameters to tune here
    },
    'Support Vector Regressor': {
        'kernel': ['linear', 'rbf'],
        'C': [0.1, 1, 10],
        # Add more hyperparameters to tune here
    },
    'Perceptron': {
        'penalty': [None, 'l2', 'l1', 'elasticnet'],
        'alpha': [0.0001, 0.001, 0.01],
        # Add more hyperparameters to tune here
    },
    'Neural Network': {
        'hidden_layer_sizes': [(50,), (100, 50), (100, 100, 50)],
        'activation': ['relu', 'tanh'],
        'alpha': [0.0001, 0.001, 0.01],
        # Add more hyperparameters to tune here
    }
    # Add parameter grids for other models
}



# Initialize a dictionary to store the best models and their scores
best_models = {}
best_scores = {}

# Iterate through the models and perform grid search
for model_name, model in models.items():
    param_grid = param_grids[model_name]
    grid_search = GridSearchCV(model, param_grid, cv=5, scoring='neg_mean_squared_error')
    grid_search.fit(X_train, y_train)
    
    # Get the best model and its score
    best_model = grid_search.best_estimator_
    best_score = -grid_search.best_score_  # Negative MSE
    
    # Store the best model and score
    best_models[model_name] = best_model
    best_scores[model_name] = best_score

# Display the results
results_df = pd.DataFrame({'Model': list(best_models.keys()), 'Best MSE Score': list(best_scores.values())})
results_df.sort_values(by='Best MSE Score', ascending=True, inplace=True)
print(results_df)
